Set-up login-screen.

+++ HeyYou/NetworkController.m
@@ -109,7 +109,7 @@
     request.HTTPMethod = @"POST";
     NSData *dotJSONData = [dot parseDotIntoJSON];
     NSUInteger length = dotJSONData.length;
-    [request setValue:[NSString stringWithFormat:@"%li", length] forHTTPHeaderField:@"Content-Length"];
+    [request setValue:[NSString stringWithFormat:@"%li", (unsigned long)length] forHTTPHeaderField:@"Content-Length"];
     [request setValue:@"application/json" forHTTPHeaderField:@"Content-Type"];
     request.HTTPBody = dotJSONData;
     NSURLSessionDataTask *dataTask = [[NSURLSession sharedSession] dataTaskWithRequest:request completionHandler:^(NSData *data, NSURLResponse *response, NSError *error) {
